backend/requirements.txt
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.3
openai==1.10.0
pillow==10.2.0
python-magic-bin==0.4.14
reportlab==4.0.9
redis==5.0.1
pydantic-settings==2.1.0
pymupdf==1.23.8
